+++ simpleble/src_c/adapter.cpp
@@ -4,6 +4,8 @@
 
 #include <cstring>
 
+bool simpleble_adapter_is_bluetooth_enabled(void) { return SimpleBLE::Safe::Adapter::bluetooth_enabled(); }
+
 size_t simpleble_adapter_get_count(void) {
     return SimpleBLE::Safe::Adapter::get_adapters().value_or(std::vector<SimpleBLE::Safe::Adapter>()).size();
 }