+++ utils/build_lib.ps1
@@ -1,6 +1,8 @@
 ﻿param (
     [Alias('c')]
     [switch]$clean = $false,
+    [Alias('e')]
+    [switch]$examples = $false,
     [validateset('x86','x64')]
     [Alias('a')]
     [string]$arch = "x64" # x86, x64
@@ -9,7 +11,11 @@
 [string]$PROJECT_ROOT = Resolve-Path $($PSScriptRoot + "\..")
 [string]$SOURCE_PATH = $PROJECT_ROOT + "\simpleble"
 [string]$BUILD_PATH = $PROJECT_ROOT + "\build_simpleble"
+[string]$INSTALL_PATH = $BUILD_PATH + "\install"
 [string]$TARGET = "Release"
+
+[string]$EXAMPLE_BUILD_PATH = $PROJECT_ROOT + "\build_simpleble_examples"
+[string]$EXAMPLE_SOURCE_PATH = $PROJECT_ROOT + "\examples\simpleble"
 
 # Validate the received architecture
 switch -regex ($arch) {
@@ -21,7 +27,6 @@ switch -regex ($arch) {
     }
 }
 
-
 # Clean directories if needed and recreate the necessary folder structure.
 if ($clean) {
     Remove-Item -Path "$BUILD_PATH" -Force -Recurse -ErrorAction SilentlyContinue
@@ -31,7 +36,17 @@ New-Item -ItemType Directory -Force -Path "$BUILD_PATH" | Out-Null
 # Run CMake to create our build files.
 cmake -S "$SOURCE_PATH" -B "$BUILD_PATH" -DCMAKE_SYSTEM_VERSION="10.0.22000.0"
 cmake --build "$BUILD_PATH" --config $TARGET --parallel 7
+cmake --install "$BUILD_PATH" --prefix "$INSTALL_PATH"
 
-#Copy all generated files to the bin folder for consistency and remove the output folder.
-Copy-item -Force -Recurse "$BUILD_PATH\bin\$TARGET\*" -Destination "$BUILD_PATH\bin\"
-#Remove-Item -Path "$PROJECT_ROOT\bin\$TARGET" -Force -Recurse -ErrorAction SilentlyContinue
+# Build the examples if needed.
+if ($examples) {
+    # Clean directories if needed and recreate the necessary folder structure.
+    if ($clean) {
+        Remove-Item -Path "$EXAMPLE_BUILD_PATH" -Force -Recurse -ErrorAction SilentlyContinue
+    }
+    New-Item -ItemType Directory -Force -Path "$EXAMPLE_BUILD_PATH" | Out-Null
+
+    # Run CMake to create our build files.
+    cmake -S "$EXAMPLE_SOURCE_PATH" -B "$EXAMPLE_BUILD_PATH" -DCMAKE_SYSTEM_VERSION="10.0.22000.0" -DSIMPLEDBUS_LOCAL=ON
+    cmake --build "$EXAMPLE_BUILD_PATH" --config $TARGET --parallel 7
+}